Added 'useReadingProblems' hook to generate passages and monitor statistics. Moved onTimeComplete function to the useReadingProblems hook, and updated onTimeComplete to create statistic from metrics. One bug is that when generating next passage, the progress bar stops moving
